MQTT wip Moved USB drivers out of the Dexcom package

+++ core/src/main/java/com/nightscout/core/mqtt/MqttPinger.java
@@ -15,7 +15,7 @@ public interface MqttPinger extends MqttPingerObservable {
 
     public boolean isNetworkActive();
 
-    public void setKeepAliveInterval(long ms);
+    public void setKeepAliveInterval(int ms);
 
     public void setMqttClient(MqttClient mqttClient);
 }
